+++ database/migrations/2020_03_09_190831_create_products_table.php
@@ -15,7 +15,7 @@ class CreateProductsTable extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('store_id');
+            $table->integer('store_id');
             $table->string('sku');
             $table->string('platform_id');
             $table->string('name');
